requirements.txt
# Trading Platform
alpaca-py

# Importing API keys
keyring

# Parsing US Congress Trading Data
requests
pandas
lxml
beautifulsoup4